#22 - renamed route config files for clarity
