+++ app/server/apps/users/routes.js
@@ -1,12 +1,12 @@
 'use strict';
 var auth = require('../../middleware/authenticate'),
     routeGenerator = require('../../route-generator'),
-    names = require('./names'),
+    methods = require('./methods'),
     middlewares = require('./middlewares');
 
 module.exports = {
     pre: {
         all: [auth.isAuthenticated]
     },
-    routes: routeGenerator(names, middlewares)
+    routes: routeGenerator(methods, middlewares)
 };
